put graphdb in separate folder; prepare for qlever support
