+++ internal/synchronizer/client_test.go
@@ -7,14 +7,15 @@ import (
 	"context"
 	"io"
 	"net/http"
+	"path/filepath"
 	"strings"
 	"testing"
 	"time"
 
 	"github.com/internetofwater/nabu/internal/common"
 	"github.com/internetofwater/nabu/internal/crawl"
+	"github.com/internetofwater/nabu/internal/synchronizer/graphdb"
 	"github.com/internetofwater/nabu/internal/synchronizer/s3"
-	"github.com/internetofwater/nabu/internal/synchronizer/triplestore"
 
 	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
@@ -62,7 +63,7 @@ type SynchronizerClientSuite struct {
 	// minio container that gleaner will send data to
 	minioContainer s3.MinioContainer
 	// graphdb container that nabu will sync with
-	graphdbContainer triplestore.GraphDBContainer
+	graphdbContainer graphdb.GraphDBContainer
 }
 
 func (suite *SynchronizerClientSuite) SetupSuite() {
@@ -88,7 +89,7 @@ func (suite *SynchronizerClientSuite) SetupSuite() {
 	err = suite.minioContainer.ClientWrapper.MakeDefaultBucket()
 	require.NoError(t, err)
 
-	graphdbContainer, err := triplestore.NewGraphDBContainer("iow", "./triplestore/testdata/iow-config.ttl")
+	graphdbContainer, err := graphdb.NewGraphDBContainer("iow", filepath.Join("graphdb", "testdata", "iow-config.ttl"))
 	suite.Require().NoError(err)
 	suite.graphdbContainer = graphdbContainer
