+++ internal/synchronizer/graphdb/client_test.go
@@ -1,7 +1,7 @@
 // Copyright 2025 Lincoln Institute of Land Policy
 // SPDX-License-Identifier: Apache-2.0
 
-package triplestore
+package graphdb
 
 import (
 	"context"
