+++ internal/synchronizer/graphdb/testdata/iow-config.ttl
